+++ MTGSearch/DBSearchViewController.m
@@ -112,7 +112,7 @@
     
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
     
-    [app_delegate trackEventWithCategory:kUACategoryUI andAction:kUAActionClick andLabel:[NSString stringWithFormat:@"card_at_pos:%d", indexPath.row]];
+    [app_delegate trackEventWithCategory:kUACategoryUI andAction:kUAActionClick andLabel:[NSString stringWithFormat:@"card_at_pos:%ld", (long)indexPath.row]];
 }
 
 - (void)searchBarSearchButtonClicked:(UISearchBar *)searchBar{
